Add RP2040 build (#530)
Fix examples to build on RP2040
Clean up generator warnings
No SPDIF on RP2040 (yet)
Add back C sources for AAC on ARM
No WiFi streaming on RP2040
No webradio on RP2040
No talking clock (NTP) on RP2040 either

+++ src/AudioGeneratorMOD.cpp
@@ -791,7 +791,7 @@ void AudioGeneratorMOD::GetSample(int16_t sample[2])
         Mixer.channelSampleNumber[channel] != FatBuffer.channelSampleNumber[channel]) {
 
       uint32_t toRead = Mixer.sampleEnd[Mixer.channelSampleNumber[channel]] - samplePointer + 1;
-      if (toRead > fatBufferSize) toRead  = fatBufferSize;
+      if (toRead > (uint32_t)fatBufferSize) toRead  = fatBufferSize;
 
       if (!file->seek(samplePointer, SEEK_SET)) {
         stop();
